+++ lib/apytypes/_array_functions.py
@@ -2450,6 +2450,59 @@ def import_csv(
     return arr
 
 
+def meshgrid(
+    *arrays: APyArray,
+    indexing: Literal["xy", "ij"] = "xy",
+) -> list[APyArray]:
+    """
+    Create a list of coordinate arrays based on several coordinate vectors (1-D arrays).
+    The input vectors must all have the same bit specifiers.
+
+    .. versionadded:: 0.4
+
+    Parameters
+    ----------
+    arrays : :class:`APyFixedArray`, :class:`APyFloatArray`, :class:`APyCFixedArray`, :class:`APyCFloatArray`
+        The coordinate vectors (1-D arrays).
+    indexing : {"xy", "ij"}, default: "xy"
+        The indexing mode to use for the meshgrid. "xy" is the standard Cartesian indexing,
+        while "ij" is the matrix indexing.
+
+    Returns
+    -------
+    result : list of :class:`APyFixedArray`, :class:`APyFloatArray`, :class:`APyCFixedArray`, :class:`APyCFloatArray`
+        List of 2-D arrays representing the meshgrid.
+
+    Examples
+    --------
+    >>> import apytypes as apy
+    >>> x = apy.fx(range(3), int_bits=5, frac_bits=0)
+    >>> y = apy.fx(range(4, 7), int_bits=5, frac_bits=0)
+    >>> xx, yy = apy.meshgrid(x, y)
+    >>> xx
+    APyFixedArray([[0, 1, 2],
+                   [0, 1, 2],
+                   [0, 1, 2]], int_bits=5, frac_bits=0)
+    >>> yy
+    APyFixedArray([[4, 4, 4],
+                   [5, 5, 5],
+                   [6, 6, 6]], int_bits=5, frac_bits=0)
+
+    """
+    if not arrays:
+        raise ValueError("meshgrid: at least one array is required")
+
+    if not all(type(arr) is type(arrays[0]) for arr in arrays):
+        raise ValueError("meshgrid: input arrays must be of same type")
+
+    try:
+        meshgrid = arrays[0]._meshgrid
+    except AttributeError:
+        raise ValueError("meshgrid: input arrays must have a meshgrid method")
+
+    return meshgrid(arrays, indexing)
+
+
 # =============================================================================
 # Helpers
 # =============================================================================