+++ CHANGELOG.md
@@ -16,7 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - ARM-Linux wheels.
 - Specialized overload typing signatures for array functions and utility
   functions.
-- Array creation method: `fullrange`.
+- Array creation methods: `fullrange`, `meshgrid`.
 - CSV import/export of arrays: `import_csv` and `export_csv`.
 
 ### Fixed